+++ src/module/Hooks.ts
@@ -1,7 +1,9 @@
-import { warn, error, debug } from "../midi-qol";
+import { warn, error, debug, i18n } from "../midi-qol";
 import { processpreCreateBetterRollsMessage, colorChatMessageHandler, diceSoNiceHandler, nsaMessageHandler, hideStuffHandler, chatDamageButtons, processcreateBetterRollMessage, mergeCardSoundPlayer, diceSoNiceUpdateMessge, recalcCriticalDamage, processItemCardCreation, hideRollUpdate, hideRollRender } from "./chatMesssageHandling";
 import { processUndoDamageCard } from "./GMAction";
 import { untargetDeadTokens, untargetAllTokens } from "./utils";
+import { actorAbilityRollPatching } from "./patching";
+import { configSettings } from "./settings";
 
 export let initHooks = () => {
   warn("Init Hooks processing");
@@ -41,5 +43,38 @@ export let initHooks = () => {
     colorChatMessageHandler(message, html, data);
     nsaMessageHandler(message, html, data);
     hideRollRender(message, html, data);
+  })
+
+  Hooks.on("renderItemSheet", (app, html, data) => {
+    if (configSettings.allowUseMacro) {
+      const element = html.find('input[name="data.chatFlavor"]').parent().parent();
+      const labelText = i18n("midi-qol.onUseMacroLabel");
+      const currentMacro = getProperty(app.object.data, "flags.midi-qol.onUseMacroName") ?? "";
+      const macroField = `<div class="form-group"><label>${labelText}</label><input type="text" name="flags.midi-qol.onUseMacroName" value="${currentMacro}"/> </div>`;
+      element.append(macroField)
+    }
+  })
+
+  Hooks.on('dropCanvasData', function(canvas, dropData) {
+    if (dropData.type !== "Item") return true;;
+    let grid_size = canvas.scene.data.grid
+
+    canvas.tokens.targetObjects({
+        x: dropData.x-grid_size/2,
+        y: dropData.y-grid_size/2,
+        height: grid_size,
+        width: grid_size
+    });
+
+    const actor = game.actors.get(dropData.actorId);
+    const item = actor && actor.items.get(dropData.data._id);
+    if (!actor || !item) console.error("actor / item broke ", actor, item);
+    if (item.type === "spell") {
+      //@ts-ignore
+      actor.useSpell(item, {configureDialog: true})
+    } else {
+      //@ts-ignore
+      item.roll();
+    }
   })
 }